app/requirements.txt
Flask==3.0.0
apscheduler==3.10.1
python-cinderclient==9.3.0
requests==2.31.0
keystoneauth1==5.2.0
